pyproject.toml
[project]
name = "moltbook-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-ai>=1.56.0",
    "python-dotenv>=1.2.1",
]
